move trusted setup in dedicated file
